+++ cmd/cycloid/middleware/pipeline.go
@@ -93,11 +93,6 @@ func (m *middleware) UpdatePipeline(org, project, env, pipeline, variables strin
 	}
 
 	p := resp.GetPayload()
-	// TODO this validate have been removed https://github.com/cycloidio/youdeploy-http-api/issues/2262
-	// err = p.Validate(strfmt.Default)
-	// if err != nil {
-	// 	return err
-	// }
 
 	d := p.Data
 	return d, err
@@ -307,4 +302,26 @@ func (m *middleware) TriggerPipelineBuild(org, project, env, job string) error {
 	_, err := m.api.OrganizationPipelinesJobsBuild.CreateBuild(params, common.ClientCredentials(&org))
 
 	return err
+}
+
+
+func (m *middleware) ListPipelines(org string) ([]*models.Pipeline, error) {
+
+	params := organization_pipelines.NewGetPipelinesParams()
+	params.SetOrganizationCanonical(org)
+
+	resp, err := m.api.OrganizationPipelines.GetPipelines(params, common.ClientCredentials(&org))
+	if err != nil {
+		return nil, err
+	}
+
+	p := resp.GetPayload()
+	// TODO this validate have been removed https://github.com/cycloidio/youdeploy-http-api/issues/2262
+	// err = p.Validate(strfmt.Default)
+	// if err != nil {
+	// 	return err
+	// }
+
+	d := p.Data
+	return d, err
 }